Fix PFP for tracks and showers
Shower's ID is always stored as -999 (at the art root file). To fix
this, instead of using the map between PFP and ShowerID, I use the
association and create the showerID when looping through the showers.

I also updated the selection isTrack and isShower, to deal with the fact
that both recob::Tracks and recob::Showers have all PFParticles stored.